requirements.txt
Flask==3.0.0
googlemaps==4.10.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
Pillow==10.2.0
